requirements.txt
Flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
cohere==4.38
python-dotenv==1.0.1
